refactor addmeasurement component
